+++ app/l10n/fontedit_en.ts
@@ -2,54 +2,118 @@
 <!DOCTYPE TS>
 <TS version="2.1" language="en_US">
 <context>
+    <name>AboutDialog</name>
+    <message>
+        <location filename="../ui/aboutdialog.ui" line="14"/>
+        <source>About FontEdit</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../ui/aboutdialog.ui" line="75"/>
+        <source>&lt;html&gt;&lt;head/&gt;&lt;body&gt;&lt;p&gt;&lt;span style=&quot; font-size:18pt;&quot;&gt;FontEdit &lt;/span&gt;&lt;span style=&quot; font-size:13pt; font-weight:400;&quot;&gt;v##version## build ##build##&lt;/span&gt;&lt;/p&gt;&lt;p&gt;&lt;span style=&quot; font-size:13pt; font-weight:400;&quot;&gt;Copyright ##year## Dominik Kapusta&lt;/span&gt;&lt;/p&gt;&lt;p align=&quot;center&quot;&gt;&lt;a href=&quot;https://github.com/ayoy/fontedit&quot;&gt;&lt;span style=&quot; font-size:13pt; font-weight:400; text-decoration: underline; color:#0068da;&quot;&gt;Get Source Code&lt;/span&gt;&lt;/a&gt;&lt;span style=&quot; font-size:13pt; font-weight:400;&quot;&gt;•&lt;/span&gt;&lt;a href=&quot;https://github.com/ayoy/fontedit/issues&quot;&gt;&lt;span style=&quot; font-size:13pt; font-weight:400; text-decoration: underline; color:#0068da;&quot;&gt;Report a Bug&lt;/span&gt;&lt;/a&gt;&lt;/p&gt;&lt;p&gt;&lt;span style=&quot; font-size:13pt; font-weight:400;&quot;&gt;This program is distributed under the terms of &lt;/span&gt;&lt;a href=&quot;https://www.gnu.org/licenses/gpl-3.0.en.html&quot;&gt;&lt;span style=&quot; font-size:13pt; font-weight:400; text-decoration: underline; color:#0068da;&quot;&gt;General Public License v3&lt;/span&gt;&lt;/a&gt;&lt;span style=&quot; font-size:13pt;&quot;&gt;.&lt;/span&gt;&lt;/p&gt;&lt;p&gt;&lt;span style=&quot; font-size:13pt; font-weight:400;&quot;&gt;&lt;br/&gt;Icons made by &lt;/span&gt;&lt;a href=&quot;https://www.flaticon.com/authors/smashicons&quot;&gt;&lt;span style=&quot; font-size:13pt; font-weight:400; text-decoration: underline; color:#0068da;&quot;&gt;Smashicons&lt;/span&gt;&lt;/a&gt;&lt;span style=&quot; font-size:13pt; font-weight:400;&quot;&gt; and &lt;/span&gt;&lt;a href=&quot;https://www.flaticon.com/authors/freepik&quot;&gt;&lt;span style=&quot; font-size:13pt; font-weight:400; text-decoration: underline; color:#0068da;&quot;&gt;Freepik&lt;/span&gt;&lt;/a&gt;&lt;span style=&quot; font-size:13pt; font-weight:400;&quot;&gt; from &lt;/span&gt;&lt;a href=&quot;https://www.flaticon.com/&quot;&gt;&lt;span style=&quot; font-size:13pt; font-weight:400; text-decoration: underline; color:#0068da;&quot;&gt;www.flaticon.com&lt;/span&gt;&lt;/a&gt;&lt;span style=&quot; font-size:13pt; font-weight:400;&quot;&gt;.&lt;/span&gt;&lt;a href=&quot;https://www.flaticon.com/&quot;&gt;&lt;span style=&quot; font-size:13pt; font-weight:400; text-decoration: underline; color:#0068da;&quot;&gt;&lt;br/&gt;&lt;/span&gt;&lt;/a&gt;&lt;/p&gt;&lt;/body&gt;&lt;/html&gt;</source>
+        <translation></translation>
+    </message>
+</context>
+<context>
+    <name>AddGlyphDialog</name>
+    <message>
+        <location filename="../addglyphdialog.ui" line="20"/>
+        <source>Add New Glyph</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../addglyphdialog.ui" line="29"/>
+        <source>Adding Glyph for </source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../addglyphdialog.ui" line="38"/>
+        <source>Add empty Glyph</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../addglyphdialog.ui" line="55"/>
+        <source>Initialize with font character:</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../addglyphdialog.ui" line="65"/>
+        <source>Current document font not found in system</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../addglyphdialog.ui" line="78"/>
+        <source>Type character here</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../addglyphdialog.ui" line="102"/>
+        <source>Copy existing Glyph:</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../addglyphdialog.cpp" line="15"/>
+        <source>Add a new Glyph to %1</source>
+        <translation></translation>
+    </message>
+</context>
+<context>
+    <name>GlyphInfoWidget</name>
+    <message>
+        <location filename="../ui/glyphinfowidget.cpp" line="34"/>
+        <source>Exported</source>
+        <translation></translation>
+    </message>
+</context>
+<context>
     <name>MainWindow</name>
     <message>
-        <location filename="../mainwindow.ui" line="28"/>
+        <location filename="../mainwindow.ui" line="32"/>
         <source>Edit Font</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="78"/>
+        <location filename="../mainwindow.ui" line="72"/>
         <source>Ctrl+R</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="248"/>
+        <location filename="../mainwindow.ui" line="266"/>
         <source>TextLabel</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="264"/>
+        <location filename="../mainwindow.ui" line="298"/>
         <source>Source Code</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="300"/>
+        <location filename="../mainwindow.ui" line="334"/>
         <source>Updating source code...</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="316"/>
+        <location filename="../mainwindow.ui" line="350"/>
         <source>Output Format</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="334"/>
+        <location filename="../mainwindow.ui" line="388"/>
         <source>Misc Options</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="340"/>
+        <location filename="../mainwindow.ui" line="394"/>
         <source>Invert Bits</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="347"/>
+        <location filename="../mainwindow.ui" line="401"/>
         <source>Reverse Bits (MSB)</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="411"/>
+        <location filename="../mainwindow.ui" line="465"/>
         <source>Print</source>
         <translation></translation>
     </message>
@@ -59,253 +123,352 @@
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="354"/>
+        <location filename="../mainwindow.ui" line="240"/>
+        <source>Toggle exported state by pressing Space on a selected glyph</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.ui" line="243"/>
+        <location filename="../mainwindow.ui" line="702"/>
+        <source>Show non-exported Glyphs</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.ui" line="359"/>
+        <source>Export Selected Glyphs</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.ui" line="369"/>
+        <source>Export All Glyphs</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.ui" line="408"/>
         <source>Include Line Spacing</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="364"/>
+        <location filename="../mainwindow.ui" line="418"/>
         <source>Font Array Name</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="376"/>
+        <location filename="../mainwindow.ui" line="430"/>
         <source>font</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="386"/>
+        <location filename="../mainwindow.ui" line="440"/>
         <source>Indentation</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="418"/>
+        <location filename="../mainwindow.ui" line="472"/>
         <source>Export Source Code...</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="441"/>
+        <location filename="../mainwindow.ui" line="495"/>
         <source>File</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="457"/>
+        <location filename="../mainwindow.ui" line="511"/>
         <source>Edit</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="473"/>
+        <location filename="../mainwindow.ui" line="524"/>
+        <source>Help</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.ui" line="531"/>
+        <source>View</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.ui" line="543"/>
         <source>New</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="476"/>
+        <location filename="../mainwindow.ui" line="546"/>
+        <location filename="../mainwindow.ui" line="558"/>
         <source>Ctrl+N</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="485"/>
+        <location filename="../mainwindow.ui" line="555"/>
         <source>Import Font</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="488"/>
-        <source>Meta+Ctrl+N</source>
-        <translation></translation>
-    </message>
-    <message>
-        <location filename="../mainwindow.ui" line="497"/>
-        <location filename="../mainwindow.cpp" line="454"/>
+        <location filename="../mainwindow.ui" line="567"/>
+        <location filename="../mainwindow.cpp" line="668"/>
         <source>Reset Glyph</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="506"/>
+        <location filename="../mainwindow.ui" line="576"/>
         <source>Export...</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="515"/>
-        <location filename="../mainwindow.cpp" line="336"/>
+        <location filename="../mainwindow.ui" line="585"/>
+        <location filename="../mainwindow.cpp" line="492"/>
         <source>Save</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="518"/>
+        <location filename="../mainwindow.ui" line="588"/>
         <source>Ctrl+S</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="527"/>
-        <location filename="../mainwindow.cpp" line="470"/>
+        <location filename="../mainwindow.ui" line="597"/>
+        <location filename="../mainwindow.cpp" line="684"/>
         <source>Reset Font</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="536"/>
+        <location filename="../mainwindow.ui" line="606"/>
         <source>Print...</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="539"/>
+        <location filename="../mainwindow.ui" line="609"/>
         <source>Ctrl+P</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="544"/>
+        <location filename="../mainwindow.ui" line="614"/>
         <source>Save As...</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="547"/>
+        <location filename="../mainwindow.ui" line="617"/>
         <source>Ctrl+Shift+S</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="556"/>
+        <location filename="../mainwindow.ui" line="626"/>
+        <location filename="../mainwindow.cpp" line="397"/>
         <source>Add Glyph</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="559"/>
+        <location filename="../mainwindow.ui" line="629"/>
         <source>Ctrl+Shift+N</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="564"/>
+        <location filename="../mainwindow.ui" line="634"/>
         <source>Quit</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="567"/>
+        <location filename="../mainwindow.ui" line="637"/>
         <source>Ctrl+Q</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="576"/>
+        <location filename="../mainwindow.ui" line="646"/>
         <source>Copy Glyph</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="585"/>
+        <location filename="../mainwindow.ui" line="655"/>
         <source>Paste Glyph</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="594"/>
+        <location filename="../mainwindow.ui" line="664"/>
         <source>Open...</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="597"/>
+        <location filename="../mainwindow.ui" line="667"/>
         <source>Ctrl+O</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="602"/>
+        <location filename="../mainwindow.ui" line="672"/>
         <source>Recent files</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.ui" line="607"/>
+        <location filename="../mainwindow.ui" line="677"/>
         <source>Close</source>
         <translation></translation>
     </message>
     <message>
+        <location filename="../mainwindow.ui" line="682"/>
+        <source>About FontEdit</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.ui" line="691"/>
+        <location filename="../mainwindow.cpp" line="415"/>
+        <location filename="../mainwindow.cpp" line="425"/>
+        <source>Delete Glyph</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.ui" line="707"/>
+        <source>Check for Updates...</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="204"/>
+        <source>Check for updates at start-up</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="213"/>
+        <source>Update Available</source>
+        <translation></translation>
+    </message>
+    <message>
         <location filename="../mainwindow.cpp" line="214"/>
+        <source>FontEdit %1 is available (you have %2).
+Get the new version from GitHub Releases Page.</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="218"/>
+        <source>Go to Releases Page</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="219"/>
+        <source>Dismiss</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="228"/>
+        <source>No Update Available</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="229"/>
+        <source>You&apos;re using the latest version of FontEdit</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="300"/>
         <source>Start by importing a system font or opening an existing document</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="217"/>
+        <location filename="../mainwindow.cpp" line="303"/>
         <source>Select a glyph on the right to edit it</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="220"/>
+        <location filename="../mainwindow.cpp" line="306"/>
         <source>Click and drag to paint, hold Alt or Ctrl to erase.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="255"/>
+        <location filename="../mainwindow.cpp" line="347"/>
         <source>Select Font</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="274"/>
+        <location filename="../mainwindow.cpp" line="366"/>
         <source>Open Document</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="306"/>
+        <location filename="../mainwindow.cpp" line="416"/>
+        <source>You can only delete Glyphs at the end of the list. This Glyph will be cleared instead of deleted (to ensure that other Glyphs&apos; indexes are unchanged).</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="425"/>
+        <source>Clear Glyph</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="462"/>
         <source>Save Document</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="314"/>
+        <location filename="../mainwindow.cpp" line="470"/>
         <source>Error</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="336"/>
+        <location filename="../mainwindow.cpp" line="492"/>
         <source>Don&apos;t Save</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="336"/>
+        <location filename="../mainwindow.cpp" line="492"/>
         <source>Cancel</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="339"/>
+        <location filename="../mainwindow.cpp" line="495"/>
         <source>Do you want to save the changes you made? Your changes will be lost if you don&apos;t save them.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="374"/>
+        <location filename="../mainwindow.cpp" line="543"/>
+        <source>Toggle Glyph Exported</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../mainwindow.cpp" line="589"/>
         <source>Size (full): %1x%2px</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="375"/>
+        <location filename="../mainwindow.cpp" line="590"/>
         <source>Size (adjusted): %1x%2px</source>
         <translation></translation>
     </message>
     <message numerus="yes">
-        <location filename="../mainwindow.cpp" line="376"/>
+        <location filename="../mainwindow.cpp" line="591"/>
         <source>%n Glyph(s)</source>
         <translation>
             <numerusform>%n Glyph</numerusform>
             <numerusform>%n Glyphs</numerusform>
         </translation>
     </message>
+    <message numerus="yes">
+        <location filename="../mainwindow.cpp" line="592"/>
+        <source>%n to export</source>
+        <translation>
+            <numerusform>%n to export</numerusform>
+            <numerusform>%n to export</numerusform>
+        </translation>
+    </message>
     <message>
-        <location filename="../mainwindow.cpp" line="419"/>
+        <location filename="../mainwindow.cpp" line="635"/>
         <source>Edit Glyph</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="441"/>
-        <source>Switch Active Glyph</source>
-        <translation></translation>
-    </message>
-    <message>
-        <location filename="../mainwindow.cpp" line="469"/>
+        <location filename="../mainwindow.cpp" line="683"/>
         <source>Are you sure you want to reset all changes to the font? This operation cannot be undone.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="523"/>
+        <location filename="../mainwindow.cpp" line="746"/>
         <source>Save Source Code</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="538"/>
+        <location filename="../mainwindow.cpp" line="761"/>
         <source>Source code successfully exported.</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindow.cpp" line="540"/>
+        <location filename="../mainwindow.cpp" line="763"/>
         <source>Unable to write to file: </source>
         <translation></translation>
     </message>
@@ -313,12 +476,12 @@
 <context>
     <name>MainWindowModel</name>
     <message>
-        <location filename="../mainwindowmodel.cpp" line="59"/>
+        <location filename="../mainwindowmodel.cpp" line="68"/>
         <source>Tab</source>
         <translation></translation>
     </message>
     <message numerus="yes">
-        <location filename="../mainwindowmodel.cpp" line="61"/>
+        <location filename="../mainwindowmodel.cpp" line="70"/>
         <source>%n Space(s)</source>
         <translation>
             <numerusform>%n Space</numerusform>
@@ -326,13 +489,26 @@
         </translation>
     </message>
     <message>
-        <location filename="../mainwindowmodel.cpp" line="140"/>
+        <location filename="../mainwindowmodel.cpp" line="145"/>
         <source>New Document</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../mainwindowmodel.cpp" line="146"/>
+        <location filename="../mainwindowmodel.cpp" line="151"/>
         <source>Edited</source>
+        <translation></translation>
+    </message>
+</context>
+<context>
+    <name>QObject</name>
+    <message>
+        <location filename="../command.h" line="39"/>
+        <source>Switch Active Glyph</source>
+        <translation></translation>
+    </message>
+    <message>
+        <location filename="../fontfaceviewmodel.cpp" line="90"/>
+        <source>Custom font</source>
         <translation></translation>
     </message>
 </context>