+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## 1.1.0
+## 1.1.0 - 2020-05-31
 * Export subset of font characters (only the characters that you really
   need in the font) - helps to reduce font size when you only use a bunch
   of characters and not the whole ASCII table
 * Exported source code is wrapped at around 80 columns
+* Added pseudocode in the generated source code, explaining how to retrieve
+  individual font glyphs
 * More user-friendly Undo/Redo functionality
 * Added checking for updates on-demand from the menu and automatically
   at start-up (the latter can be disabled)